+++ soc/arm/ti_lm3s6965/CMakeLists.txt
@@ -1,4 +1,6 @@
 zephyr_sources(
   soc.c
   soc_config.c
+  reboot.S
+  sys_arch_reboot.c
   )
